Fixed Jest test
